requirements.txt
streamlit
numpy
opencv-python
Pillow
tensorflow
keras
matplotlib